+++ examples/benchmark/bm_pico_kd_tree.cpp
@@ -1,4 +1,4 @@
-#include <pico_adaptor.hpp>
+#include <pico_toolshed/pico_adaptor.hpp>
 #include <pico_tree/kd_tree.hpp>
 
 #include "benchmark.hpp"
