+++ examples/benchmark/format_ascii.hpp
@@ -2,7 +2,6 @@
 
 #include <iomanip>
 #include <pico_tree/core.hpp>
-#include <point.hpp>
 
 namespace pico_tree {
